2.x hotfix: add messaging in install for edge cases such as missing files, session driver should be database

+++ src/Console/InstallCommand.php
@@ -99,8 +99,23 @@ class InstallCommand extends Command
         }
 
         $this->replaceInFile("'SESSION_DRIVER', 'file'", "'SESSION_DRIVER', 'database'", config_path('session.php'));
-        $this->replaceInFile('SESSION_DRIVER=file', 'SESSION_DRIVER=database', base_path('.env'));
-        $this->replaceInFile('SESSION_DRIVER=file', 'SESSION_DRIVER=database', base_path('.env.example'));
+
+        collect(['.env', '.env.example', '.env.local'])->map(function ($envFile) {
+            $envFileBasePath = base_path($envFile);
+
+            if (file_exists($envFileBasePath)) {
+                $this->replaceInFile('SESSION_DRIVER=file', 'SESSION_DRIVER=database', $envFileBasePath);
+
+                return $envFile;
+            }
+        })->filter()->whenEmpty(function () {
+            $this->info(<<<'END_OF_INFO'
+                \nJetstream attempted to update .env, .env.local, and .env.example to ensure the session driver is
+                set to "database", but could not find one. Jetstream works best with database sessions, for more
+                information see: https://jetstream.laravel.com/2.x/features/browser-sessions.html
+                END_OF_INFO
+            );
+        });
     }
 
     /**
@@ -684,6 +699,12 @@ EOF;
      */
     protected function replaceInFile($search, $replace, $path)
     {
+        if (! file_exists($path)) {
+            $this->warn("Attempting to update {$path} but the file was not found");
+
+            return;
+        }
+
         file_put_contents($path, str_replace($search, $replace, file_get_contents($path)));
     }
